requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==4.2.20
djangorestframework==3.16.0
hiredis==3.1.0
pika==1.3.2
psycopg==3.2.6
psycopg-binary==3.2.6
redis==6.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
